Crawling: use same User-Agent header during `robots.txt` retrieval as recipe webpage retrieval
